+++ Classes/Commands/AppleScript.h
@@ -17,8 +17,10 @@
 @interface LowerVolumeCommand : NSScriptCommand {} @end
 @interface FullVolumeCommand : NSScriptCommand {} @end
 @interface MuteCommand : NSScriptCommand {} @end
+@interface UnmuteCommand : NSScriptCommand {} @end
 @interface TiredCommand : NSScriptCommand {} @end
 
 @interface NSApplication (HermesScripting)
 - (NSNumber*) volume;
+- (void) setVolume: (NSNumber*) volume;
 @end